# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Legacy Link Backend API",
    "version": "0.6.0",
    "description": "API Legacy Link: token đọc cho GET, token ghi cho POST; timestamp đo là epoch milliseconds, ID DB là chuỗi. C16 bổ sung profile version, preview, lịch sử thao tác PostgreSQL, metrics, freshness. Xem deploy/README.md và docs/SO-TAY-HIEU-DU-AN-LEGACY-LINK.md. Rate limit mỗi socket IP/process: 300 GET và 30 thao tác/phút; 429 kèm Retry-After. Thêm deliveryHealth và diagnostics.delivery để quan sát queue ESP32; không cần endpoint riêng. AI: POST /ai/chat và /ai/query chỉ đọc, chấp nhận token đọc; Gemini chỉ chọn tool. UNDERHEAT dùng lowAlarm riêng, yêu cầu đơn vị nhiệt độ."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/ai/chat": {
      "post": {
        "summary": "Câu hỏi tiếng Việt: Gemini chọn công cụ; backend xác minh dữ liệu",
        "description": "Chấp nhận read hoặc write bearer token. Chỉ gửi question hoặc action. Context giữ ID kết quả 10 phút, tối đa 200 phiên/process. Model tối đa 5 câu/phút/process, 1 đồng thời; 1 request/câu, không retry. Timeout provider 12s, tổng tools 8s. Không có key/hết quota thì fallback minh bạch. Không thực thi SQL, URL hoặc điều khiển máy.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "conversationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "question": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "overheat",
                      "underheat",
                      "offline",
                      "summary",
                      "alerts"
                    ]
                  }
                },
                "oneOf": [
                  {
                    "required": [
                      "question"
                    ]
                  },
                  {
                    "required": [
                      "action"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dữ liệu từ database; mode=rules hoặc gemini; không bịa số đo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotResponse"
                }
              }
            }
          },
          "400": {
            "description": "Sai tham số"
          },
          "401": {
            "description": "Thiếu hoặc sai token"
          },
          "403": {
            "description": "Origin/phiên không được phép"
          },
          "404": {
            "description": "Máy không tồn tại"
          },
          "429": {
            "description": "Rate limit HTTP"
          },
          "503": {
            "description": "Nguồn dữ liệu không khả dụng"
          }
        }
      }
    },
    "/ai/query": {
      "post": {
        "summary": "AI Copilot chỉ đọc: quick action không dùng Gemini",
        "description": "Chấp nhận read hoặc write bearer token. Chỉ gửi question hoặc action. Context giữ ID kết quả 10 phút, tối đa 200 phiên/process. Model tối đa 5 câu/phút/process, 1 đồng thời; 1 request/câu, không retry. Timeout provider 12s, tổng tools 8s. Không có key/hết quota thì fallback minh bạch. Không thực thi SQL, URL hoặc điều khiển máy.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "conversationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "overheat",
                      "underheat",
                      "offline",
                      "summary",
                      "alerts"
                    ]
                  }
                },
                "required": [
                  "action"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dữ liệu từ database; mode=rules hoặc gemini; không bịa số đo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotResponse"
                }
              }
            }
          },
          "400": {
            "description": "Sai tham số"
          },
          "401": {
            "description": "Thiếu hoặc sai token"
          },
          "403": {
            "description": "Origin/phiên không được phép"
          },
          "404": {
            "description": "Máy không tồn tại"
          },
          "429": {
            "description": "Rate limit HTTP"
          },
          "503": {
            "description": "Nguồn dữ liệu không khả dụng"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Alarm": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "code": {
            "type": "string",
            "enum": [
              "OVERHEAT",
              "OVERCURRENT",
              "OVERSPEED",
              "VIBRATION",
              "UNDERHEAT"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "value": {
            "type": "number",
            "nullable": true
          },
          "metricKey": {
            "type": "string",
            "nullable": true
          },
          "eventId": {
            "type": "string",
            "nullable": true
          },
          "acknowledgedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "deviceId",
          "timestamp",
          "code",
          "severity"
        ]
      },
...
      "CopilotResponse": {
        "type": "object",
        "required": [
          "conversationId",
          "mode",
          "text",
          "notice",
          "results"
        ],
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "rules",
              "gemini"
            ]
          },
          "text": {
            "type": "string"
          },
          "notice": {
            "type": "string",
            "nullable": true
          },
          "results": {
            "type": "array",
            "maxItems": 4,
            "items": {
              "type": "object",
              "required": [
                "tool",
                "source",
                "queriedAt"
              ],
              "properties": {
                "tool": {
                  "type": "string"
                },
                "source": {
                  "type": "string",
                  "enum": [
                    "database"
                  ]
                },
                "queriedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "truncated": {
                  "type": "boolean"
                },
                "note": {
                  "type": "string"
                },
                "devices": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "alerts": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alarm"
                  }
                },
                "series": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token đọc hoặc ghi do người vận hành cấp; POST yêu cầu token ghi."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}